// EmployeeManager.ClientApp/src/app/core/api/generated/model/employeeCreateDto.ts
/**
 * EmployeeManager.WebAPI
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { Gender } from './gender';


export interface EmployeeCreateDto { 
    firstName?: string | null;
    lastName?: string | null;
    middleName?: string | null;
    birthDate?: string;
    gender?: Gender;
    email?: string | null;
    phoneNumber?: string | null;
    joinedDate?: string;
    superiorId?: number | null;
    addressId?: number | null;
    jobCategoryIds?: Array<number> | null;
}
export namespace EmployeeCreateDto {
}